download_artifact returns path list; add convert_to_argo method to Workflow
